@@ -1,6 +1,6 @@
 {
   "name": "storybook-vue3-router",
-  "version": "6.0.2",
+  "version": "7.0.0",
   "description": "A Storybook decorator that allows you to build stories for your routing-aware components.",
   "keywords": [
     "storybook-addons",
@@ -14,8 +14,7 @@
   "author": "Nick McBurney",
   "license": "MIT",
   "type": "module",
-  "main": "dist/index.cjs",
-  "module": "dist/index.mjs",
+  "main": "dist/index.js",
   "typings": "dist/index.d.ts",
   "types": "dist/index.d.ts",
   "files": [
@@ -33,31 +32,31 @@
     "test": "npx playwright test"
   },
   "dependencies": {
-    "vue-router": "^4.5.1"
+    "vue-router": "^4.6.3"
   },
   "devDependencies": {
-    "@playwright/test": "^1.54.1",
-    "@storybook/addon-docs": "^9.0.18",
-    "@storybook/vue3-vite": "^9.0.18",
-    "@types/node": "^18.19.120",
-    "@vitejs/plugin-vue": "^4.6.2",
+    "@playwright/test": "^1.56.1",
+    "@storybook/addon-docs": "^10.0.7",
+    "@storybook/vue3-vite": "^10.0.7",
+    "@types/node": "^20.19.25",
+    "@vitejs/plugin-vue": "^6.0.1",
     "concurrently": "^6.5.1",
     "esbuild": "^0.17.19",
     "prettier": "^2.8.8",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "rimraf": "^3.0.2",
+    "rimraf": "^6.1.0",
     "rollup": "^3.29.5",
     "rollup-plugin-dts": "^5.3.1",
     "rollup-plugin-esbuild": "^5.0.0",
-    "storybook": "^9.0.18",
-    "typescript": "^5.8.3",
-    "vite": "^7.0.6",
-    "vue": "^3.5.18",
+    "storybook": "^10.0.7",
+    "typescript": "^5.9.3",
+    "vite": "^7.2.2",
+    "vue": "^3.5.24",
     "zx": "^1.15.2"
   },
   "peerDependencies": {
-    "@storybook/vue3": "9.x",
+    "@storybook/vue3": "10.x",
     "vue": "3.x",
     "vue-router": "4.x"
   },
